% Auto-generated: do not edit by hand
\name{dscApp}

\alias{dscApp}

\title{App component}

\description{
Component description
}

\usage{
dscApp(id=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

