% Auto-generated: do not edit by hand
\name{dscDashSolanaComponents}

\alias{dscDashSolanaComponents}

\title{DashSolanaComponents component}

\description{

}

\usage{
dscDashSolanaComponents(id=NULL, network=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{network}{A value equal to: 'devnet', 'mainnet', 'testnet'. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

