% Auto-generated: do not edit by hand
\name{dscSendSol}

\alias{dscSendSol}

\title{SendSol component}

\description{

}

\usage{
dscSendSol(id=NULL, className=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{className}{Character. Adds CSS class name(s).}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

