% Auto-generated: do not edit by hand
\name{dscSolanaWalletButtonWrapper}

\alias{dscSolanaWalletButtonWrapper}

\title{SolanaWalletButtonWrapper component}

\description{

}

\usage{
dscSolanaWalletButtonWrapper(children=NULL, id=NULL, className=NULL, rpcEndpoint=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The child components to be rendered inside the wrapper.}

\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{className}{Character. Adds CSS class name(s).}

\item{rpcEndpoint}{Character. The custom RPC endpoint for the wallet.
If it's not provided, the default endpoint for the specified network will be used.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

