% Auto-generated: do not edit by hand
\name{dscSolanaWalletMultiButton}

\alias{dscSolanaWalletMultiButton}

\title{SolanaWalletMultiButton component}

\description{

}

\usage{
dscSolanaWalletMultiButton(id=NULL, initialPublicKeyState=NULL, network=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{initialPublicKeyState}{Character. }

\item{network}{A value equal to: 'devnet', 'mainnet', 'testnet'. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

