% Auto-generated: do not edit by hand
\name{dscTransactionButtonWrapper}

\alias{dscTransactionButtonWrapper}

\title{TransactionButtonWrapper component}

\description{

}

\usage{
dscTransactionButtonWrapper(children=NULL, id=NULL, className=NULL,
onTransactionSent=NULL, transactionInstructions=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{className}{Character. Adds CSS class name(s).}

\item{onTransactionSent}{}

\item{transactionInstructions}{List of characters. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

