% Auto-generated: do not edit by hand
\name{dscWalletContextProvider}

\alias{dscWalletContextProvider}

\title{WalletContextProvider component}

\description{

}

\usage{
dscWalletContextProvider(children=NULL, id=NULL, className=NULL, network=NULL,
rpcEndpoint=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{className}{Character. Adds CSS class name(s).}

\item{network}{A value equal to: 'mainnet-beta', 'testnet', 'devnet'. }

\item{rpcEndpoint}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

