% Auto-generated: do not edit by hand
\name{dscDashSolanaComponents}

\alias{dscDashSolanaComponents}

\title{DashSolanaComponents component}

\description{
Component description
}

\usage{
dscDashSolanaComponents(id=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

