% Auto-generated: do not edit by hand
\name{dscSolanaWalletMultiButton}

\alias{dscSolanaWalletMultiButton}

\title{SolanaWalletMultiButton component}

\description{

}

\usage{
dscSolanaWalletMultiButton(id=NULL, network=NULL, publicKeyState=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{network}{A value equal to: 'devnet', 'mainnet', 'testnet'. }

\item{publicKeyState}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

